made some progress with understanding transfer learning
It is really easy. The one thing is, I will need to figure out a better way to extract any layer I want, because the tutorial only did the last layer.

But I assume it's as easy as picking the correct name of the layer to override and then making more forward passes over whatever I define.

Will be testing tomorrow.