some questions I need to ask:
(1) how exactly to compute the 3-channel mean and std?
(2) how to visualize it after loading? I ran into a number of problems, and the closest I got it working was when I actually ignored 2 of the dimensions and did a PIL save, which gave the correct colors. Yet, even that single-channel thing resulted in a yellow-ish color for the cv2.imwrite, very confusing
(3) and of course I also need to fix the transforms in the firs place, we shouldn't be cropping this heavily because I can't even see if it's successful or not. But that stuff can wait until (1) and (2) are done.

Write up a gist and ask on the PyTorch forums for advice on visualizing the data, and make sure to describe all the critical steps and I also need to give them data.